convert report_auto
